changed matrix from double to float modifications on fraction constructor implemented the ability to add decimal numbers on the model

+++ include/Matrix.hpp
@@ -26,7 +26,7 @@ public:
     Matrix();
     Matrix(unsigned rows, unsigned columns);
     Matrix(Matrix &m);
-    Matrix(const std::vector<std::vector<double>> &m);
+    Matrix(const std::vector<std::vector<float>> &m);
     
     //Destructor
     ~Matrix();
